requirements.txt
streamlit
pandas
pypdf
python-docx
pillow
pytesseract
pdf2image
python-dotenv
pymysql
